salvage MessageQueue provided by jdeqsim and used in other places
