+++ matsim/src/main/java/org/matsim/core/mobsim/messagequeue/Scheduler.java
@@ -1,11 +1,9 @@
-
 /* *********************************************************************** *
  * project: org.matsim.*
- * MessageQueueEngine.java
  *                                                                         *
  * *********************************************************************** *
  *                                                                         *
- * copyright       : (C) 2019 by the members listed in the COPYING,        *
+ * copyright       : (C) 2008 by the members listed in the COPYING,        *
  *                   LICENSE and WARRANTY file.                            *
  * email           : info at matsim dot org                                *
  *                                                                         *
@@ -19,26 +17,27 @@
  *                                                                         *
  * *********************************************************************** */
 
- package org.matsim.core.mobsim.qsim.messagequeueengine;
+package org.matsim.core.mobsim.messagequeue;
 
-import org.matsim.core.mobsim.framework.events.MobsimBeforeSimStepEvent;
-import org.matsim.core.mobsim.framework.listeners.MobsimBeforeSimStepListener;
-import org.matsim.core.mobsim.messagequeue.SteppableScheduler;
+/**
+ * The scheduler of the micro-simulation.
+ *
+ * @author rashid_waraich
+ */
+public class Scheduler {
 
-import jakarta.inject.Inject;
+	protected final MessageQueue queue;
 
-class MessageQueueEngine implements MobsimBeforeSimStepListener {
-
-	private final SteppableScheduler scheduler;
-
-	@Inject
-	MessageQueueEngine(final SteppableScheduler scheduler) {
-		this.scheduler = scheduler;
+	public Scheduler(MessageQueue queue) {
+		this(queue, Double.MAX_VALUE);
 	}
 
-	@Override
-	public void notifyMobsimBeforeSimStep(MobsimBeforeSimStepEvent e) {
-		scheduler.doSimStep(e.getSimulationTime());
+	public Scheduler(MessageQueue messageQueue, double simulationEndTime) {
+		this.queue = messageQueue;
+	}
+
+	public void schedule(Message m) {
+		queue.putMessage(m);
 	}
 
 }
