+++ matsim/src/main/java/org/matsim/core/mobsim/qsim/messagequeueengine/MessageQueueModule.java
@@ -1,4 +1,3 @@
-
 /* *********************************************************************** *
  * project: org.matsim.*
  * MessageQueueModule.java
@@ -19,11 +18,11 @@
  *                                                                         *
  * *********************************************************************** */
 
- package org.matsim.core.mobsim.qsim.messagequeueengine;
+package org.matsim.core.mobsim.qsim.messagequeueengine;
 
-import org.matsim.core.mobsim.jdeqsim.MessageQueue;
+import org.matsim.core.mobsim.messagequeue.MessageQueue;
 import org.matsim.core.mobsim.qsim.AbstractQSimModule;
-import org.matsim.core.mobsim.qsim.jdeqsimengine.SteppableScheduler;
+import org.matsim.core.mobsim.messagequeue.SteppableScheduler;
 
 public class MessageQueueModule extends AbstractQSimModule {
 	static public final String COMPONENT_NAME = "MessageQueueEngine";
@@ -33,8 +32,7 @@ public class MessageQueueModule extends AbstractQSimModule {
 		bind(MessageQueue.class).asEagerSingleton();
 		bind(SteppableScheduler.class).asEagerSingleton();
 		bind(MessageQueueEngine.class).asEagerSingleton();
-		
-//		addNamedComponent(MessageQueueEngine.class, COMPONENT_NAME);
+
 		this.addQSimComponentBinding( COMPONENT_NAME ).to( MessageQueueEngine.class ) ;
 
 	}
