+++ matsim/src/main/java/org/matsim/withinday/package-info.java
@@ -1,4 +1,3 @@
-
 /* *********************************************************************** *
  * project: org.matsim.*
  * package-info.java
@@ -21,8 +20,7 @@
 
  /**
  * This package contains some experimental code which adds within day replanning features to the
- * QSim. The JDEQSim is not supported at the moment and currently there no plans to change that
- * (if you think you *really* need that, feel free to implement it...).
+ * QSim.
  * </p>
  * A tutorial which shows the usage of this code should be available by the end of 2011.
  * </p>
